Change Poisson to return float for lambtha. Add Exponential class

+++ math/0x03-probability/poisson.py
@@ -36,7 +36,7 @@ class Poisson():
             elif len(data) < 2:
                 raise ValueError('data must contain multiple values')
             else:
-                self.lambtha = (sum(data)) / (len(data))
+                self.lambtha = float(sum(data) / len(data))
 
     def pmf(self, k):
         """
